Reuse float-converted hidden_states in T5LayerNorm

+++ src/transformers/models/t5/modeling_t5.py
@@ -245,7 +245,8 @@ class T5LayerNorm(nn.Module):
 
     def forward(self, hidden_states):
         # layer norm should always be calculated in float32
-        variance = hidden_states.to(torch.float32).pow(2).mean(-1, keepdim=True)
+        hidden_states = hidden_states.to(torch.float32)
+        variance = hidden_states.pow(2).mean(-1, keepdim=True)
         hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
 
         # convert into half-precision if necessary
